code/sql-python/python/requirements.txt
mysql-connector-python
psycopg2
sqlalchemy
